requirements.txt
flask>=1.0
flask-cors==3.0.8
flask-restplus==0.12.1
google-api-core==1.1.1
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
web3>=4.9.2
Werkzeug==0.15.3
requests>=2.22.0
Pillow==6.2.0
python-dateutil>=2.8.0
